refactor search engine
