+++ PowerPointLabs/PowerPointLabs/ImageSearch/SearchEngine/GoogleEngine.Const.cs
@@ -1,0 +1,9 @@
+﻿namespace PowerPointLabs.ImageSearch.SearchEngine
+{
+    partial class GoogleEngine
+    {
+        public const int NumOfItemsPerSearch = 30;
+        public const int NumOfItemsPerRequest = 10;
+        public const int MaxNumOfItems = 100;
+    }
+}
