+++ PowerPointLabs/PowerPointLabs/ImageSearch/SearchEngine/GoogleEngine.Events.cs
@@ -1,0 +1,38 @@
+﻿using PowerPointLabs.ImageSearch.SearchEngine.VO;
+using RestSharp;
+
+namespace PowerPointLabs.ImageSearch.SearchEngine
+{
+    partial class GoogleEngine
+    {
+        public delegate void WhenFailEventDelegate(IRestResponse response);
+
+        private event WhenFailEventDelegate WhenFailDelegate;
+
+        public GoogleEngine WhenFail(WhenFailEventDelegate action)
+        {
+            WhenFailDelegate += action;
+            return this;
+        }
+
+        public delegate void WhenSucceedEventDelegate(GoogleSearchResults results, int startIdx);
+
+        private event WhenSucceedEventDelegate WhenSucceedDelegate;
+
+        public GoogleEngine WhenSucceed(WhenSucceedEventDelegate action)
+        {
+            WhenSucceedDelegate += action;
+            return this;
+        }
+
+        public delegate void WhenCompletedEventDelegate(bool isSuccessful);
+
+        private event WhenCompletedEventDelegate WhenCompletedDelegate;
+
+        public GoogleEngine WhenCompleted(WhenCompletedEventDelegate action)
+        {
+            WhenCompletedDelegate += action;
+            return this;
+        }
+    }
+}
